split stuff out
