+++ src/www/app/app.ts
@@ -20,7 +20,8 @@ import {ServerListItem, ServerConnectionState} from '../views/servers_view';
 import {Clipboard} from './clipboard';
 import {EnvironmentVariables} from './environment';
 import {OutlineErrorReporter} from './error_reporter';
-import {OutlineServer, OutlineServerRepository} from './outline_server';
+import {OutlineServer} from './outline_server';
+import {OutlineServerRepository} from './outline_server_repository';
 import {Settings, SettingsKey} from './settings';
 import {Updater} from './updater';
 import {UrlInterceptor} from './url_interceptor';
