+++ src/www/app/main.ts
@@ -19,7 +19,8 @@ import {EventQueue} from '../model/events';
 import {App} from './app';
 import {onceEnvVars} from './environment';
 import {NativeNetworking} from './net';
-import {OutlineServerRepository, shadowsocksConfigToAccessKey} from './outline_server';
+import {shadowsocksConfigToAccessKey} from './outline_server_access_key';
+import {OutlineServerRepository} from './outline_server_repository';
 import {OutlinePlatform} from './platform';
 import {Settings} from './settings';
 import {TunnelFactory} from './tunnel';
